Disable building SSL files due to problems
As of now we don't use the SSL-functionality in Folly, and it fails
to build on some platforms with OpenSSL 1.1.1b we want to use to
get TLS 1.3. Given that we don't use it let's just ignore the
failing code and hope that it is fixed upstream when we want to
start using it.